requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
autopep8==1.7.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cli-helpers==2.2.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.6
cryptography==36.0.2
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-ratelimit==3.0.1
django-storages==1.10.1
djangorestframework==3.12.2
executing==1.0.0
happybase==1.2.0
idna==3.4
importlib-resources==5.9.0
ipython==8.5.0
jedi==0.18.1
jsonfield==3.1.0
kombu==5.2.4
matplotlib-inline==0.1.6
mycli==1.26.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pycodestyle==2.9.1
pycparser==2.21
pycrypto==2.6.1
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
python-memcached==1.59
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlglot==6.2.1
sqlparse==0.4.3
stack-data==0.5.0
swapper==1.3.0
tabulate==0.8.10
thriftpy2==0.4.14
toml==0.10.2
traitlets==5.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
